+++ src/libaudtag/audtag.h
@@ -35,7 +35,7 @@ enum
 void tag_set_verbose (bool_t verbose);
 
 bool_t tag_tuple_read (Tuple * tuple, VFSFile *fd);
-bool_t tag_image_read (VFSFile * handle, void * * data, int * size);
+bool_t tag_image_read (VFSFile * handle, void * * data, int64_t * size);
 
 /* new_type specifies the type of tag (see the TAG_TYPE_* enum) that should be
  * written if the file does not have any existing tag. */
